Refactor: 서비스 로직, Review의 별점 Review : reviewRating 타입을 double => int로 변경 AuctionService : 필요없는 @Transactional 제거, 내부 메서드 접근 제한자 public으로 변경 TicketService : 필요없는 @Transactional 제거, 내부 메서드 접근 제한자 public으로 변경, 수수료 계산하는 로직에 조건 추가 TicketServiceTest : 수수료 계산 테스트

+++ src/main/java/shootingstar/var/dto/req/ReviewSaveReqDto.java
@@ -16,5 +16,5 @@ public class ReviewSaveReqDto {
 
     @Min(value = 1)
     @Max(value = 5)
-    private double reviewRating;
+    private int reviewRating;
 }
